+++ src/test/java/com/github/ciselab/lampion/guided/metric/metrics/F1_ScoreTest.java
@@ -1,11 +1,9 @@
-package com.github.ciselab.metric.metrics;
+package com.github.ciselab.lampion.guided.metric.metrics;
 
 import com.github.ciselab.lampion.guided.algorithms.MetamorphicIndividual;
 import com.github.ciselab.lampion.guided.configuration.Configuration;
 import com.github.ciselab.lampion.guided.metric.Metric;
-import com.github.ciselab.lampion.guided.metric.metrics.Precision;
 import com.github.ciselab.lampion.guided.metric.metrics.F1;
-import com.github.ciselab.lampion.guided.metric.metrics.Precision;
 import com.github.ciselab.lampion.guided.metric.metrics.Recall;
 import com.github.ciselab.lampion.guided.support.GenotypeSupport;
 import com.github.ciselab.lampion.guided.support.MetricCache;
@@ -14,27 +12,33 @@ import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-public class PrecisionTest {
+public class F1_ScoreTest {
+
+    @Test
+    public void checkNameTest() {
+        F1 metric = new F1();
+        assertEquals("F1", metric.getName());
+    }
 
     @Test
     public void testCanBeBiggerThanOne_shouldNotBe(){
-        Metric metric = new Precision();
+        Metric metric = new F1();
         assertFalse(metric.canBeBiggerThanOne());
     }
 
     @Test
     public void testIsSecondary_shouldNotBe(){
-        Metric metric = new Precision();
+        Metric metric = new F1();
         assertFalse(metric.isSecondary());
     }
 
     @Tag("File")
     @Test
-    public void testPrecision_PathIsOk_shouldGiveValue(){
+    public void testF1Score_PathIsOk_shouldGiveValue(){
         var config = new Configuration();
         MetricCache cache = new MetricCache();
         GenotypeSupport support = new GenotypeSupport(cache,config);
-        Precision metric = new Precision();
+        F1 metric = new F1();
 
         MetamorphicIndividual testObject = new MetamorphicIndividual(support, 0);
         testObject.setResultPath("./src/test/resources/metric_files");
@@ -47,11 +51,11 @@ public class PrecisionTest {
     }
 
     @Test
-    public void testPrecision_PathIsBad_NaNValue(){
+    public void testF1_PathIsBad_NaNValue(){
         var config = new Configuration();
         MetricCache cache = new MetricCache();
         GenotypeSupport support = new GenotypeSupport(cache,config);
-        Precision metric = new Precision();
+        F1 metric = new F1();
 
         MetamorphicIndividual testObject = new MetamorphicIndividual(support, 0);
         testObject.setResultPath("./src/test/bad_path");
@@ -65,55 +69,55 @@ public class PrecisionTest {
 
     @Test
     public void testEquality_isEqualToItself(){
-        Metric metric = new Precision();
+        Metric metric = new F1();
 
         assertEquals(metric,metric);
     }
 
     @Test
     public void testEquality_sameWeight_isEqual(){
-        Metric a = new Precision();
-        Metric b = new Precision();
+        Metric a = new F1();
+        Metric b = new F1();
 
         assertEquals(a,b);
     }
 
     @Test
     public void testEquality_differentWeight_notEqual(){
-        Metric a = new Precision();
+        Metric a = new F1();
         a.setWeight(0.75);
-        Metric b = new Precision();
+        Metric b = new F1();
         b.setWeight(0.5);
         assertNotEquals(a,b);
     }
 
     @Test
     public void testEquality_againstNonMetric_isNotEqual(){
-        Metric metric = new Precision();
+        Metric metric = new F1();
         Double other = 5.0;
         assertNotEquals(other,metric);
     }
 
     @Test
     public void testHashCode_isEqualToItself(){
-        Metric metric = new Precision();
+        Metric metric = new F1();
 
         assertEquals(metric.hashCode(),metric.hashCode());
     }
 
     @Test
     public void testHashCode_sameWeight_isEqual(){
-        Metric a = new Precision();
-        Metric b = new Precision();
+        Metric a = new F1();
+        Metric b = new F1();
 
         assertEquals(a.hashCode(),b.hashCode());
     }
 
     @Test
     public void testHashCode_differentWeight_notEqual(){
-        Metric a = new Precision();
+        Metric a = new F1();
         a.setWeight(0.75);
-        Metric b = new Precision();
+        Metric b = new F1();
         b.setWeight(0.5);
         assertNotEquals(a.hashCode(),b.hashCode());
     }
