+++ src/test/java/com/github/ciselab/lampion/guided/metric/metrics/InputLengthTest.java
@@ -1,4 +1,4 @@
-package com.github.ciselab.metric.metrics;
+package com.github.ciselab.lampion.guided.metric.metrics;
 
 import com.github.ciselab.lampion.guided.metric.Metric;
 import com.github.ciselab.lampion.guided.metric.metrics.PredictionLength;
