+++ src/test/java/com/github/ciselab/lampion/guided/metric/metrics/MRRTest.java
@@ -1,10 +1,10 @@
-package com.github.ciselab.metric.metrics;
+package com.github.ciselab.lampion.guided.metric.metrics;
 
 import com.github.ciselab.lampion.guided.algorithms.MetamorphicIndividual;
 import com.github.ciselab.lampion.guided.configuration.Configuration;
 import com.github.ciselab.lampion.guided.metric.Metric;
-import com.github.ciselab.lampion.guided.metric.metrics.F1;
-import com.github.ciselab.lampion.guided.metric.metrics.Recall;
+import com.github.ciselab.lampion.guided.metric.metrics.*;
+import com.github.ciselab.lampion.guided.metric.metrics.MRR;
 import com.github.ciselab.lampion.guided.support.GenotypeSupport;
 import com.github.ciselab.lampion.guided.support.MetricCache;
 import org.junit.jupiter.api.Tag;
@@ -12,33 +12,27 @@ import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-public class F1_ScoreTest {
-
-    @Test
-    public void checkNameTest() {
-        F1 metric = new F1();
-        assertEquals("F1", metric.getName());
-    }
+public class MRRTest {
 
     @Test
     public void testCanBeBiggerThanOne_shouldNotBe(){
-        Metric metric = new F1();
+        Metric metric = new PercentageMRR();
         assertFalse(metric.canBeBiggerThanOne());
     }
 
     @Test
     public void testIsSecondary_shouldNotBe(){
-        Metric metric = new F1();
+        Metric metric = new PercentageMRR();
         assertFalse(metric.isSecondary());
     }
 
     @Tag("File")
     @Test
-    public void testF1Score_PathIsOk_shouldGiveValue(){
+    public void testMRRScore_PathIsOk_shouldGiveValue(){
         var config = new Configuration();
         MetricCache cache = new MetricCache();
         GenotypeSupport support = new GenotypeSupport(cache,config);
-        F1 metric = new F1();
+        MRR metric = new MRR();
 
         MetamorphicIndividual testObject = new MetamorphicIndividual(support, 0);
         testObject.setResultPath("./src/test/resources/metric_files");
@@ -51,11 +45,11 @@ public class F1_ScoreTest {
     }
 
     @Test
-    public void testF1_PathIsBad_NaNValue(){
+    public void testMRR_PathIsBad_NaNValue(){
         var config = new Configuration();
         MetricCache cache = new MetricCache();
         GenotypeSupport support = new GenotypeSupport(cache,config);
-        F1 metric = new F1();
+        MRR metric = new MRR();
 
         MetamorphicIndividual testObject = new MetamorphicIndividual(support, 0);
         testObject.setResultPath("./src/test/bad_path");
@@ -69,55 +63,55 @@ public class F1_ScoreTest {
 
     @Test
     public void testEquality_isEqualToItself(){
-        Metric metric = new F1();
+        Metric metric = new MRR();
 
         assertEquals(metric,metric);
     }
 
     @Test
     public void testEquality_sameWeight_isEqual(){
-        Metric a = new F1();
-        Metric b = new F1();
+        Metric a = new MRR();
+        Metric b = new MRR();
 
         assertEquals(a,b);
     }
 
     @Test
     public void testEquality_differentWeight_notEqual(){
-        Metric a = new F1();
+        Metric a = new MRR();
         a.setWeight(0.75);
-        Metric b = new F1();
+        Metric b = new MRR();
         b.setWeight(0.5);
         assertNotEquals(a,b);
     }
 
     @Test
     public void testEquality_againstNonMetric_isNotEqual(){
-        Metric metric = new F1();
+        Metric metric = new MRR();
         Double other = 5.0;
         assertNotEquals(other,metric);
     }
 
     @Test
     public void testHashCode_isEqualToItself(){
-        Metric metric = new F1();
+        Metric metric = new MRR();
 
         assertEquals(metric.hashCode(),metric.hashCode());
     }
 
     @Test
     public void testHashCode_sameWeight_isEqual(){
-        Metric a = new F1();
-        Metric b = new F1();
+        Metric a = new MRR();
+        Metric b = new MRR();
 
         assertEquals(a.hashCode(),b.hashCode());
     }
 
     @Test
     public void testHashCode_differentWeight_notEqual(){
-        Metric a = new F1();
+        Metric a = new MRR();
         a.setWeight(0.75);
-        Metric b = new F1();
+        Metric b = new MRR();
         b.setWeight(0.5);
         assertNotEquals(a.hashCode(),b.hashCode());
     }
