+++ src/test/java/com/github/ciselab/lampion/guided/metric/metrics/Percentage_MRRTest.java
@@ -1,11 +1,12 @@
-package com.github.ciselab.metric.metrics;
+package com.github.ciselab.lampion.guided.metric.metrics;
 
 import com.github.ciselab.lampion.guided.algorithms.MetamorphicIndividual;
 import com.github.ciselab.lampion.guided.configuration.Configuration;
 import com.github.ciselab.lampion.guided.metric.Metric;
+import com.github.ciselab.lampion.guided.metric.metrics.PercentageMRR;
+import com.github.ciselab.lampion.guided.metric.metrics.F1;
+import com.github.ciselab.lampion.guided.metric.metrics.PercentageMRR;
 import com.github.ciselab.lampion.guided.metric.metrics.Recall;
-import com.github.ciselab.lampion.guided.metric.metrics.Recall;
-import com.github.ciselab.lampion.guided.metric.metrics.Transformations;
 import com.github.ciselab.lampion.guided.support.GenotypeSupport;
 import com.github.ciselab.lampion.guided.support.MetricCache;
 import org.junit.jupiter.api.Tag;
@@ -13,28 +14,27 @@ import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-public class RecallTest {
+public class Percentage_MRRTest {
 
     @Test
     public void testCanBeBiggerThanOne_shouldNotBe(){
-        Metric metric = new Recall();
+        Metric metric = new PercentageMRR();
         assertFalse(metric.canBeBiggerThanOne());
     }
 
     @Test
     public void testIsSecondary_shouldNotBe(){
-        Metric metric = new Recall();
+        Metric metric = new PercentageMRR();
         assertFalse(metric.isSecondary());
     }
 
-
     @Tag("File")
     @Test
-    public void testF1Score_PathIsOk_shouldGiveValue(){
+    public void testPMRR_PathIsOk_shouldGiveValue(){
         var config = new Configuration();
         MetricCache cache = new MetricCache();
         GenotypeSupport support = new GenotypeSupport(cache,config);
-        Recall metric = new Recall();
+        PercentageMRR metric = new PercentageMRR();
 
         MetamorphicIndividual testObject = new MetamorphicIndividual(support, 0);
         testObject.setResultPath("./src/test/resources/metric_files");
@@ -47,11 +47,11 @@ public class RecallTest {
     }
 
     @Test
-    public void testF1_PathIsBad_NaNValue(){
+    public void testPMRR_PathIsBad_NaNValue(){
         var config = new Configuration();
         MetricCache cache = new MetricCache();
         GenotypeSupport support = new GenotypeSupport(cache,config);
-        Recall metric = new Recall();
+        PercentageMRR metric = new PercentageMRR();
 
         MetamorphicIndividual testObject = new MetamorphicIndividual(support, 0);
         testObject.setResultPath("./src/test/bad_path");
@@ -62,58 +62,57 @@ public class RecallTest {
         assertEquals(Double.NaN,result);
     }
 
-
     @Test
     public void testEquality_isEqualToItself(){
-        Metric metric = new Recall();
+        Metric metric = new PercentageMRR();
 
         assertEquals(metric,metric);
     }
 
     @Test
     public void testEquality_sameWeight_isEqual(){
-        Metric a = new Recall();
-        Metric b = new Recall();
+        Metric a = new PercentageMRR();
+        Metric b = new PercentageMRR();
 
         assertEquals(a,b);
     }
 
     @Test
     public void testEquality_differentWeight_notEqual(){
-        Metric a = new Recall();
+        Metric a = new PercentageMRR();
         a.setWeight(0.75);
-        Metric b = new Recall();
+        Metric b = new PercentageMRR();
         b.setWeight(0.5);
         assertNotEquals(a,b);
     }
 
     @Test
     public void testEquality_againstNonMetric_isNotEqual(){
-        Metric metric = new Recall();
+        Metric metric = new PercentageMRR();
         Double other = 5.0;
         assertNotEquals(other,metric);
     }
 
     @Test
     public void testHashCode_isEqualToItself(){
-        Metric metric = new Recall();
+        Metric metric = new PercentageMRR();
 
         assertEquals(metric.hashCode(),metric.hashCode());
     }
 
     @Test
     public void testHashCode_sameWeight_isEqual(){
-        Metric a = new Recall();
-        Metric b = new Recall();
+        Metric a = new PercentageMRR();
+        Metric b = new PercentageMRR();
 
         assertEquals(a.hashCode(),b.hashCode());
     }
 
     @Test
     public void testHashCode_differentWeight_notEqual(){
-        Metric a = new Recall();
+        Metric a = new PercentageMRR();
         a.setWeight(0.75);
-        Metric b = new Recall();
+        Metric b = new PercentageMRR();
         b.setWeight(0.5);
         assertNotEquals(a.hashCode(),b.hashCode());
     }
